+++ src/Spotflow.InMemory.Azure.EventHubs/Hooks/ConsumerEventHubScope.cs
@@ -1,4 +1,8 @@
+
 namespace Spotflow.InMemory.Azure.EventHubs.Hooks;
 
-public record ConsumerEventHubScope(string EventHubNamespaceName, string EventHubName, string ConsumerGroup, string PartitionId)
-    : EventHubScope(EventHubNamespaceName, EventHubName);
+public record ConsumerEventHubScope(string EventHubNamespaceName, string EventHubName, string ConsumerGroup)
+    : EventHubScope(EventHubNamespaceName, EventHubName)
+{
+    public PartitionConsumerEventHubScope WithPartition(string partitionId) => new(EventHubNamespaceName, EventHubName, ConsumerGroup, partitionId);
+}