+++ src/Spotflow.InMemory.Azure.EventHubs/Hooks/ConsumerOperations.cs
@@ -5,6 +5,8 @@ public enum ConsumerOperations
 {
     None = 0,
     ReceiveBatch = 1,
-    All = ReceiveBatch
+    GetEventHubProperties = 2,
+    GetPartitionProperties = 4,
+    All = ReceiveBatch | GetEventHubProperties | GetPartitionProperties
 }