+++ src/Spotflow.InMemory.Azure.EventHubs/Hooks/Contexts/GetEventHubPropertiesConsumerBeforeHookContext.cs
@@ -1,0 +1,4 @@
+namespace Spotflow.InMemory.Azure.EventHubs.Hooks.Contexts;
+
+public class GetEventHubPropertiesConsumerBeforeHookContext(ConsumerEventHubScope scope, InMemoryEventHubProvider provider, CancellationToken cancellationToken)
+    : ConsumerBeforeHookContext(scope, ConsumerOperations.GetEventHubProperties, provider, cancellationToken);