+++ src/Spotflow.InMemory.Azure.EventHubs/Hooks/Contexts/GetPartitionPropertiesConsumerBeforeHookContext.cs
@@ -1,0 +1,4 @@
+namespace Spotflow.InMemory.Azure.EventHubs.Hooks.Contexts;
+
+public class GetPartitionPropertiesConsumerBeforeHookContext(PartitionConsumerEventHubScope scope, InMemoryEventHubProvider provider, CancellationToken cancellationToken)
+    : PartitionConsumerBeforeHookContext(scope, ConsumerOperations.GetPartitionProperties, provider, cancellationToken);