+++ src/Spotflow.InMemory.Azure.EventHubs/Hooks/Contexts/PartitionConsumerAfterHookContext.cs
@@ -1,0 +1,9 @@
+using Spotflow.InMemory.Azure.EventHubs.Hooks.Internals;
+
+namespace Spotflow.InMemory.Azure.EventHubs.Hooks.Contexts;
+
+public abstract class PartitionConsumerAfterHookContext(PartitionConsumerBeforeHookContext before)
+    : ConsumerAfterHookContext(before), IConsumerPartitionOperation
+{
+    public string PartitionId => before.PartitionId;
+}