+++ src/Spotflow.InMemory.Azure.EventHubs/Hooks/Contexts/ReceiveBatchBeforeHookContext.cs
@@ -1,5 +1,4 @@
 namespace Spotflow.InMemory.Azure.EventHubs.Hooks.Contexts;
 
-public class ReceiveBatchBeforeHookContext(ConsumerEventHubScope scope, InMemoryEventHubProvider provider, CancellationToken cancellationToken)
-    : ConsumerBeforeHookContext(scope, ConsumerOperations.ReceiveBatch, provider, cancellationToken);
-
+public class ReceiveBatchBeforeHookContext(PartitionConsumerEventHubScope scope, InMemoryEventHubProvider provider, CancellationToken cancellationToken)
+    : PartitionConsumerBeforeHookContext(scope, ConsumerOperations.ReceiveBatch, provider, cancellationToken);