+++ src/Spotflow.InMemory.Azure.EventHubs/Hooks/Internals/IConsumerOperation.cs
@@ -2,9 +2,6 @@ namespace Spotflow.InMemory.Azure.EventHubs.Hooks.Internals;
 
 internal interface IConsumerOperation
 {
-
     string ConsumerGroup { get; }
-    string PartitionId { get; }
     ConsumerOperations Operation { get; }
 }
-