+++ src/Spotflow.InMemory.Azure.EventHubs/Hooks/Internals/IConsumerPartitionOperation.cs
@@ -1,0 +1,7 @@
+namespace Spotflow.InMemory.Azure.EventHubs.Hooks.Internals;
+
+internal interface IConsumerPartitionOperation : IConsumerOperation
+{
+
+    string PartitionId { get; }
+}