+++ src/Spotflow.InMemory.Azure.EventHubs/Hooks/PartitionConsumerEventHubScope.cs
@@ -1,0 +1,4 @@
+namespace Spotflow.InMemory.Azure.EventHubs.Hooks;
+
+public record PartitionConsumerEventHubScope(string EventHubNamespaceName, string EventHubName, string ConsumerGroup, string PartitionId)
+    : ConsumerEventHubScope(EventHubNamespaceName, EventHubName, ConsumerGroup);